commit méthodes des services en places et testées sur main, apparition sur la base

+++ src/main/java/fr/formation/models/Medecin.java
@@ -34,7 +34,7 @@ public class Medecin {
 		return idMedecin;
 	}
 
-	public void setIdMedecin(int idMedecin) {
+	public void setIdMedecin(long idMedecin) {
 		this.idMedecin = idMedecin;
 	}
